+++ movekit/__init__.py
@@ -1,5 +1,8 @@
 from .io.csv import *
 from .io.excel import *
 from .features.absolute import *
-from .preprocessing.preprocess import *
 from .preprocessing.interpolation import *
+from .preprocessing.plot import *
+from .preprocessing.preprocess import *
+from .preprocessing.resample import *
+from .preprocessing.segmentation import *
